Fix non-trivial warnings hidden by -isystem
These warnings were hidden by -isystem.

Bug: 31751828
Test: m -j checkbuild
Merged-In: Ibb1aa07a6827fcbd29fc15291fcd90280c0f608d
(cherry picked from commit 6f51c153ea5e4187a313e47ac8bad9a1a999d4e1)

+++ media/libstagefright/httplive/LiveSession.cpp
@@ -1535,7 +1535,7 @@ void LiveSession::changeConfiguration(
                     mOrigBandwidthIndex, mCurBandwidthIndex);
         }
     }
-    CHECK_LT(mCurBandwidthIndex, mBandwidthItems.size());
+    CHECK_LT((size_t)mCurBandwidthIndex, mBandwidthItems.size());
     const BandwidthItem &item = mBandwidthItems.itemAt(mCurBandwidthIndex);
 
     uint32_t streamMask = 0; // streams that should be fetched by the new fetcher
